+++ src/features/common/components/Form/index.ts
@@ -3,3 +3,4 @@ export { Error } from "./Error";
 export { Textarea } from "./Textarea";
 export { Label } from "./Label";
 export { FormInput } from "./FormInput";
+export { Datetime } from "./Datetime";
